separate more drawing from game logic
- not amazing but can use for next steps
